+++ EsportSimulator/Assets/Scripts/Data/PlayerData.cs
@@ -10,15 +10,15 @@ public class PlayerData : ScriptableObject
 {
     #region Default Variables
 
-    [SerializeField] private float defaultMoney;
+    [SerializeField] private int defaultMoney;
 
     #endregion
 
     #region Progress
 
-    [SerializeField] private float money;
-    [SerializeField] private float rating;
-    [SerializeField] private float fame;
+    [SerializeField] private int money;
+    [SerializeField] private int rating;
+    [SerializeField] private int fame;
     [SerializeField] private float workExperience;
     [SerializeField] private int workLevel;
     [SerializeField] private int houseLevel;
@@ -27,9 +27,9 @@ public class PlayerData : ScriptableObject
 
     #region Needs
 
-    [SerializeField] private float tiredness;
-    [SerializeField] private float hunger;
-    [SerializeField] private float thirst;
+    [SerializeField] private int tiredness;
+    [SerializeField] private int hunger;
+    [SerializeField] private int thirst;
 
     #endregion
 
@@ -61,30 +61,30 @@ public class PlayerData : ScriptableObject
 
     #region Getters & Setters
 
-    public float GetDefaultMoney { get { return defaultMoney; } }
-    public float GetMoney { get { return money; } }
-    public float GetRating { get { return rating; } }
-    public float GetFame { get { return fame; } }
+    public int GetDefaultMoney { get { return defaultMoney; } }
+    public int GetMoney { get { return money; } }
+    public int GetRating { get { return rating; } }
+    public int GetFame { get { return fame; } }
     public float GetWorkExperience { get { return workExperience; } }
     public int GetWorkLevel { get { return workLevel; } }
     public int GetHouseLevel { get { return houseLevel; } }
-    public float GetTiredness { get { return tiredness; } }
-    public float GetHunger { get { return hunger; } }
-    public float GetThirst { get { return thirst; } }
+    public int GetTiredness { get { return tiredness; } }
+    public int GetHunger { get { return hunger; } }
+    public int GetThirst { get { return thirst; } }
     public int GetGameKnowledge { get { return gameKnowledge; } }
     public int GetTeamPlay { get { return teamPlay; } }
     public int GetMechanics { get { return mechanics; } }
 
-    public float SetDefaultMoney { set { defaultMoney = value; } }
-    public float SetMoney { set { money = value; } }
-    public float SetRating { set { rating = value; } }
-    public float SetFame { set { fame = value; } }
+    public int SetDefaultMoney { set { defaultMoney = value; } }
+    public int SetMoney { set { money = value; } }
+    public int SetRating { set { rating = value; } }
+    public int SetFame { set { fame = value; } }
     public float SetWorkExperience { set { workExperience = value; } }
     public int SetWorkLevel { set { workLevel = value; } }
     public int SetHouseLevel { set { houseLevel = value; } }
-    public float SetTiredness { set { tiredness = value; } }
-    public float SetHunger { set { hunger = value; } }
-    public float SetThirst { set { thirst = value; } }
+    public int SetTiredness { set { tiredness = value; } }
+    public int SetHunger { set { hunger = value; } }
+    public int SetThirst { set { thirst = value; } }
     public int SetGameKnowledge { set { gameKnowledge = value; } }
     public int SetTeamPlay { set { teamPlay = value; } }
     public int SetMechanics { set { mechanics = value; } }
